Updated to 1.0.2, fixing 64-bit warnings

+++ BPXLUUIDHandler/BPXLUUIDHandler.m
@@ -161,7 +161,7 @@ static CFMutableDictionaryRef CreateKeychainQueryDictionary(BOOL legacyItem)
 	
 	if (status != noErr)
 	{
-		NSLog(@"BPXLUUIDHandler Keychain Save Error: %ld", status);
+		NSLog(@"BPXLUUIDHandler Keychain Save Error: %d", (int)status);
 		uuid = nil;
 	}
 	
@@ -219,7 +219,7 @@ static NSString *_uuid = nil;
 		else // Any other error, log it and return nil
 		{
 			CFRelease(query);
-			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %ld", status);
+			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %d", (int)status);
 			return nil;
 		}
 	}
@@ -242,7 +242,7 @@ static NSString *_uuid = nil;
 		}
 		else // Any other error, log it and return nil
 		{
-			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %ld", status);
+			NSLog(@"BPXLUUIDHandler Unhandled Keychain Error %d", (int)status);
 			return nil;
 		}
 	}
@@ -266,7 +266,7 @@ static NSString *_uuid = nil;
 		}
 		else
 		{
-			NSLog(@"BPXLUUIDHandler Could not update UUID error %ld", deleteStatus);
+			NSLog(@"BPXLUUIDHandler Could not update UUID error %d", (int)deleteStatus);
 		}
 	}
 	CFRelease(query);
@@ -298,7 +298,7 @@ static NSString *_uuid = nil;
 	status = SecItemDelete(query);
 	if (status != noErr)
 	{
-		NSLog(@"BPXLUUIDHandler Keychain Delete Error: %ld", status);
+		NSLog(@"BPXLUUIDHandler Keychain Delete Error: %d", (int)status);
 	}
 	CFRelease(query);
 }
